+++ spec/default_spec.rb
@@ -1,0 +1,12 @@
+require 'spec_helper'
+
+describe 'nfs::default' do
+  let(:chef_run) do
+    ChefSpec::Runner.new.converge(described_recipe)
+  end
+
+  it 'should include nfs::_common' do
+    expect(chef_run).to include_recipe('nfs::_common')
+  end
+
+end