+++ test/kitchen/cookbooks/nfs_test/metadata.rb
@@ -1,15 +1,9 @@
-maintainer       "Eric G. Wolfe"
-maintainer_email "wolfe21@marshall.edu"
-license          "Apache 2.0"
-description      "Tests NFS cookbook"
+maintainer 'Eric G. Wolfe'
+maintainer_email 'wolfe21@marshall.edu'
+license 'Apache 2.0'
+description 'Tests NFS cookbook'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          "0.0.1"
-
-#depends          "nfs"
-
-#recipe           "nfs_test::client"
-#recipe           "nfs_test::server"
-
-#%w{ centos ubuntu }.each do |os|
-#  supports os
-#end
+version '0.0.2'
+name 'nfs_test'
+depends 'nfs'
+depends 'line'